examples/adwords/adwords_appengine_demo/requirements.txt
googleads==7.0.0
